requirements-new.txt
# Core Dependencies
PyQt6>=6.0.0
PyJWT>=2.4.0
cryptography>=41.0.0
requests>=2.26.0
matplotlib>=3.4.3
argon2-cffi>=21.3.0
bleach>=4.1.0
tqdm>=4.62.3
aiohttp>=3.8.1
SQLAlchemy>=1.4.31
redis>=4.1.0
asyncio>=3.4.3
pydantic<2.0.0,>=1.6.2  # Pin pydantic version for compatibility

# AI Integration
openai>=0.27.0
tiktoken>=0.3.0  # For OpenAI token counting
tenacity>=8.0.1  # For API retry handling

# Security Tools
shodan>=1.27.0
fpdf>=1.7.2
pyOpenSSL>=23.2.0  # For SSL/TLS checking
sslyze>=5.1.3  # For detailed SSL/TLS analysis

# Web Framework
fastapi>=0.95.2  # Requires pydantic<2.0.0

# Development Tools
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.2

# Documentation
mkdocs>=1.2.3
mkdocs-material>=7.3.0

# Optional Dependencies
beautifulsoup4>=4.9.3  # For web scraping
python-nmap>=0.7.1    # For network scanning
python-whois>=0.7.3   # For domain information
